Get Rid of Factory
Files have been encoded using the base64 algorithm and are now included
directly in the phpmvc.php file.  For this reason, end-users can simply
download the file without having to worry about external dependencies.
The script can be run and will generate all the necessary files for the
proper operation of a PHPMVC app by decoding the base64-encoded strings
of the files previously located in the “factory” directory.